Finish PerfectNote 100% (Without Long Notes)

+++ Assets/Script/Game/PartyGame/CoinsReward.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class CoinsReward : MonoBehaviour {
     
-    public int beginY;
+    public float beginY;
     public bool hasFinishAnimation;
     private GameController controller;
